# backend/openapi/api-spec.yaml
openapi: 3.0.3
info:
  title: MSS Course Platform BFF API
  description: |
    Backend for Frontend API specification for the MSS Course Platform.
    This API defines all endpoints currently implemented using Supabase,
    prepared for migration to Python FastAPI BFF layer.
  version: 1.0.0
  contact:
    name: MSS Course Platform Team
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://your-app.vercel.app/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        message:
          type: string
...
    UserRole:
      type: string
      enum: [user, instructor, admin]
      description: User role in the system
    Profile:
      type: object
      required:
        - id
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          maxLength: 50
          nullable: true
          description: User's display name (max 50 characters)
        avatar_url:
          type: string
          format: uri
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        bio:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # Category schemas
...
    CommentBase:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
    Comment:
      allOf:
        - $ref: '#/components/schemas/CommentBase'
        - type: object
          required:
            - id
            - course_id
            - user_id
            - is_instructor_reply
          properties:
            id:
              type: string
              format: uuid
            course_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            is_instructor_reply:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            profiles:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                email:
                  type: string
                avatar_url:
                  type: string
                  nullable: true
                role:
                  $ref: '#/components/schemas/UserRole'

    # Statistics schemas
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user with email and password
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                    description: JWT authentication token
                  user:
                    $ref: '#/components/schemas/Profile'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (5 attempts per 5 minutes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Register a new user account
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - display_name
              properties:
                email:
                  type: string
                  format: email
                display_name:
                  type: string
                  maxLength: 50
                  minLength: 1
                  description: User's display name (required, 1-50 characters)
                password:
                  type: string
                  format: password
                  minLength: 6
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: Please check your email to verify your account
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (3 attempts per hour)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logout current user
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get current authenticated user's profile
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    put:
      tags:
        - Authentication
      summary: Update user profile
      description: Update current user's display name and bio
      operationId: updateProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  maxLength: 50
                  minLength: 1
                  description: User's display name (1-50 characters)
                bio:
                  type: string
                  maxLength: 500
                  nullable: true
                  description: User's bio (max 500 characters)
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /courses/{courseId}/comments:
    get:
      tags:
        - Comments
      summary: Get course comments
      description: Get all comments for a course
      operationId: getCourseComments
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  count:
                    type: integer
                    description: Total number of comments

    post:
      tags:
        - Comments
      summary: Post a comment
      description: Post a new comment on a course (rate limited - 5 per minute)
      operationId: addComment
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBase'
      responses:
        '201':
          description: Comment posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (5 comments per minute)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /comments/{id}:
    delete:
      tags:
        - Comments
      summary: Delete comment
      description: Delete a comment (owner or admin only)
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comment deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Category endpoints
...